// src/api/__generated__/story/model/actorNftMintDigestResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */

export interface ActorNftMintDigestResponse {
  actorCollectionId?: number;
  assetId?: string;
  walletAddress?: string;
  collectionMintAddress?: string;
  name?: string;
  nftChain?: string;
  nftTokenStandard?: string;
  initialPriceUsdc?: number;
  trust?: number;
  payToken?: string;
  currentSupply?: number;
  quantity?: number;
  basePrice?: number;
  totalPrice?: number;
  totalPriceWithSlippage?: number;
  nextPrice?: number;
  orderId?: number;
  issuedAt?: number;
  expiresAt?: number;
  canonicalPayload?: string;
  sig?: string;
}